+++ crates/jstzd/Cargo.toml
@@ -15,7 +15,6 @@ http.workspace = true
 octez = { path = "../octez" }
 regex.workspace = true
 reqwest.workspace = true
-rust-embed.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tempfile.workspace = true
