--- crates/jstzd/src/jstzd.rs
@@ -1,1 +1,0 @@
-pub const DEFAULT_RPC_ENDPOINT: &str = "localhost:8732";
