+++ crates/octez/Cargo.toml
@@ -16,8 +16,10 @@ description = "Bindings to Octez (Tezos) binaries"
 anyhow.workspace = true
 hex.workspace = true
 http.workspace = true
+jstz_crypto = {path = "../jstz_crypto"}
 regex.workspace = true
 reqwest.workspace = true
+rust-embed.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 signal-hook.workspace = true
