+++ crates/octez/src/lib.rs
@@ -2,19 +2,14 @@ use std::{path::PathBuf, process::Command};
 
 use anyhow::{anyhow, Result};
 
-mod async_node;
+pub mod r#async;
 mod client;
-mod endpoint;
 mod node;
-mod node_config;
 mod rollup;
 mod thread;
 
-pub use async_node::*;
 pub use client::*;
-pub use endpoint::*;
 pub use node::*;
-pub use node_config::*;
 pub use rollup::*;
 pub use thread::*;
